Separate clear and initialize database
